new file:   src/main/java/com/dusinski/holidaycalendar/controller/EventController.java 	modified:   src/main/java/com/dusinski/holidaycalendar/controller/MainController.java 	modified:   src/main/java/com/dusinski/holidaycalendar/enums/EventType.java 	modified:   src/main/java/com/dusinski/holidaycalendar/model/User.java 	modified:   src/main/resources/input_data.sql 	deleted:    src/main/resources/templates/accountVerified.html 	renamed:    src/main/resources/templates/addEventForm.html -> src/main/resources/templates/event/addEventForm.html 	renamed:    src/main/resources/templates/verificationError.html -> src/main/resources/templates/event/eventVerificationError.html 	new file:   src/main/resources/templates/event/eventVerified.html 	new file:   src/main/resources/templates/event/eventWaitForVerification.html 	renamed:    src/main/resources/templates/showEvents.html -> src/main/resources/templates/event/showEvents.html 	modified:   src/main/resources/templates/index.html 	modified:   src/main/resources/templates/security/login.html 	modified:   src/main/resources/templates/user/addUser.html 	modified:   src/main/resources/templates/user/showUsers.html
