+++ src/main/java/com/dusinski/holidaycalendar/enums/EventType.java
@@ -1,7 +1,7 @@
 package com.dusinski.holidaycalendar.enums;
 
 public enum EventType {
-    holiday,
-    sickness,
-    training
+    HOLIDAY,
+    SICKNESS,
+    TRAINING
 }
