+++ src/main/java/com/dusinski/holidaycalendar/model/User.java
@@ -11,6 +11,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
+//    @Id
     @Email
     private String email;
     private boolean isAdmin;
