--- src/main/resources/templates/accountVerified.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>accountVerified</title>
-</head>
-<body>
-    the account is accountVerified
-</body>
-</html>
