+++ src/main/resources/templates/event/addEventForm.html
@@ -30,10 +30,10 @@
                     <a class="nav-link" href="/">Calendar view</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/show">Show my holidays</a>
+                    <a class="nav-link" href="/event/show">Show my holidays</a>
                 </li>
                 <li class="nav-item active">
-                    <a class="nav-link" href="/addEvent">Add new holiday</a>
+                    <a class="nav-link" href="/event/addEvent">Add new holiday</a>
                 </li>
 
                 <li class="nav-item">
@@ -60,7 +60,7 @@
 
     </main><!-- /.container -->
     <div>
-        <form action="#" th:action="@{/addEvent}" th:object="${calendarEvent}" method="post">
+        <form action="#" th:action="@{/event/addEvent}" th:object="${calendarEvent}" method="post">
             <div class="form-group row">
                 <label class="col-sm-2 col-from-label">Name of the event</label>
                 <div class="col-sm-8">
