+++ src/main/resources/templates/event/eventVerificationError.html
