+++ src/main/resources/templates/event/eventVerified.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Your event is verified</title>
+</head>
+<body>
+        Your event is verified
+        <br>
+        Click <a href="/event/show">Click here</a> to go to event's list
+</body>
+</html>
