+++ src/main/resources/templates/event/eventWaitForVerification.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Your event is waiting for verification</title>
+</head>
+<body>
+Your event is waiting for YOUR verification. Please open your mail and click the link to verify it.
+<br>
+Click <a href="/event/show">Click here</a> to go to event's list
+</body>
+</html>
