+++ src/main/resources/templates/event/showEvents.html
@@ -29,10 +29,10 @@
                 <a class="nav-link" href="/">Calendar view</a>
             </li>
             <li class="nav-item active">
-                <a class="nav-link" href="/show">Show my holidays</a>
+                <a class="nav-link" href="/event/show">Show my holidays</a>
             </li>
             <li class="nav-item">
-                <a class="nav-link" href="/addEvent">Add new holiday</a>
+                <a class="nav-link" href="/event/addEvent">Add new holiday</a>
             </li>
             <li class="nav-item">
                 <a class="nav-link" href="/user/show">Show user list</a>
