+++ src/main/resources/templates/security/login.html
@@ -13,16 +13,32 @@
     <div th:if="${param.logout}">
         You have been logged out
     </div>
-    <form th:action="@{/login}" method="post">
+    <form th:action="@{/login}" method="post" >
         <div><label>Choose user name from the list : </label>
-        <select id="username" name="username" th:object="${usersList}">
-            <option th:each="user : ${usersList}" th:value="${user.email}" th:text="${user.email}"></option>
+        <select id="username" name="username" th:object="${usersList}"  onclick="myFunction()">
+            <option th:each="user : ${usersList}" th:value="${user.email}" th:text="${user.email}" th:title="${user.admin}" ></option>
 
         </select>
         </div>
         <!--<div><label>User name : <input type="text" name="username"></label></div>-->
-        <div><label>User password : <input type="password" name="password"></label></div>
+        <div><label>User password : <input id="password" type="password" name="password" style="display: none" ></label></div>
         <div><input type="submit" value ="Sign In"></div>
     </form>
+
+    <script>function myFunction(){
+        // console.debug("on beginning");
+        var selectList = document.getElementById("username");
+        var selectedValue = selectList.options[selectList.selectedIndex].title;
+        var passwordInputText=document.getElementById("password");
+        // console.debug("selectedValue.value: "+selectedValue);
+        if (selectedValue=="true"){
+            passwordInputText.style.display="block";
+
+        }
+        else
+        {passwordInputText.style.display="none";
+        }
+    }
+    </script>
 </body>
 </html>
