+++ src/main/resources/templates/user/addUser.html
@@ -33,10 +33,10 @@
                     <a class="nav-link" href="/">Calendar view</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="/show">Show my holidays</a>
+                    <a class="nav-link" href="/event/show">Show my holidays</a>
                 </li>
                 <li class="nav-item active">
-                    <a class="nav-link" href="/addEvent">Add new holiday</a>
+                    <a class="nav-link" href="/event/addEvent">Add new holiday</a>
                 </li>
 
                 <li class="nav-item">
